241029: refactor cells and refactor calculators
